requirements.txt
flask
flask-socketio
flask-jwt-extended
flask-sqlalchemy
bcrypt
cryptography
python-dotenv
eventlet
pytest
